+++ frontend/src/views/HomePage.vue
@@ -1,0 +1,10 @@
+
+<template>
+  <div>
+    <h1>PAGE CONTENT WILL GO HERE</h1>
+  </div>
+</template>
+
+<script setup lang="ts">
+//
+</script>